+++ NEWS.md
@@ -1,3 +1,8 @@
+# 2026-04-03
+
+- Remove `type` from predict
+- Fix binary classification in vignettes
+
 # 2025-11-04
 
 - Initial version